integrating unit test with ld layout
